make separate nodes for simulating depth images and orbiting the camera
