+++ yak_ros2/launch/demo.launch.py
@@ -9,6 +9,7 @@ from ament_index_python.packages import get_package_share_directory
 
 def generate_launch_description():
     ld = LaunchDescription([
+        # Core node for GPU-accelerated 3D reconstruction
         launch_ros.actions.Node(
             node_name='yak_ros2_node', package='yak_ros2', node_executable='yak_ros2_node', output='screen',
             remappings=[('input_depth_image', '/image')],
@@ -29,15 +30,27 @@ def generate_launch_description():
                          'volume_z': 192,
                          }]
             ),
+        # Simulate depth images of a given mesh file
         launch_ros.actions.Node(
             node_name='yak_ros2_image_simulator', package='yak_ros2', node_executable='yak_ros2_image_simulator', output='screen',
             parameters=[{
                          'base_frame': 'world',
-                         'orbit_speed': 1.0,
+                         'camera_frame': 'camera',
                          'framerate': 30.0,
                          'mesh': path.join(get_package_share_directory('yak_ros2'), 'demo', 'bun_on_table.ply'),
                          }]
             ),
+        # Broadcast transforms between the world frame and the camera frame that orbit the camera around the world frame
+        launch_ros.actions.Node(
+            node_name='yak_ros2_tf_broadcaster', package='yak_ros2', node_executable='yak_ros2_tf_broadcaster', output='screen',
+            parameters=[{
+                         'base_frame': 'world',
+                         'camera_frame': 'camera',
+                         'orbit_speed': 1.0,
+                         'hz': 30.0,
+                         }]
+            ),
+        # Fixed transform between the world frame and the origin of the TSDF volume
         launch_ros.actions.Node(
             node_name='static_tf_publisher', package='tf2_ros', node_executable='static_transform_publisher', output='screen',
             arguments=['-0.3', '-0.3', '-0.01', '0', '0', '0', '1', 'world', 'tsdf_origin'],
