+++ yak_ros2/src/demo/sim_depth_image_pub.cpp
@@ -6,7 +6,8 @@
 #include <image_transport/image_transport.h>
 #include <cv_bridge/cv_bridge.h>
 #include <tf2/transform_datatypes.h>
-#include <tf2_ros/transform_broadcaster.h>
+#include <tf2_ros/transform_listener.h>
+#include <tf2_ros/buffer.h>
 #include <tf2_eigen/tf2_eigen.h>
 
 #include <opencv2/highgui.hpp>
@@ -14,20 +15,6 @@
 #include <chrono>
 
 const static std::string DEFAULT_IMAGE_TOPIC = "image";
-
-static Eigen::Isometry3d lookat(const Eigen::Vector3d& origin, const Eigen::Vector3d& eye, const Eigen::Vector3d& up)
-{
-  Eigen::Vector3d z = (eye - origin).normalized();
-  Eigen::Vector3d x = z.cross(up).normalized();
-  Eigen::Vector3d y = z.cross(x).normalized();
-
-  auto p = Eigen::Isometry3d::Identity();
-  p.translation() = origin;
-  p.matrix().col(0).head<3>() = x;
-  p.matrix().col(1).head<3>() = y;
-  p.matrix().col(2).head<3>() = z;
-  return p;
-}
 
 int main(int argc, char** argv)
 {
@@ -37,17 +24,15 @@ int main(int argc, char** argv)
   // Setup ROS interfaces
   image_transport::ImageTransport it(node);
   image_transport::Publisher image_pub = it.advertise(DEFAULT_IMAGE_TOPIC, 1);
-  tf2_ros::TransformBroadcaster broadcaster(node);
+  tf2_ros::Buffer buffer(node->get_clock());
+  tf2_ros::TransformListener tf_listener(buffer);
 
   // declare parameters
   node->declare_parameter("mesh");
-  node->declare_parameter("world");
-  node->declare_parameter("camera");
-  node->declare_parameter("orbit_speed");
+  node->declare_parameter("base_frame");
+  node->declare_parameter("camera_frame");
   node->declare_parameter("framerate");
-  node->declare_parameter("radius");
   node->declare_parameter("focal_length");
-  node->declare_parameter("z");
   node->declare_parameter("width");
   node->declare_parameter("height");
 
@@ -65,17 +50,8 @@ int main(int argc, char** argv)
   std::string camera_frame;
   node->get_parameter_or<std::string>("camera_frame", camera_frame, "camera");
 
-  double orbit_speed;
-  node->get_parameter_or<double>("orbit_speed", orbit_speed, 1.0);
-
   double framerate;
   node->get_parameter_or<double>("framerate", framerate, 30.0);
-
-  double radius;
-  node->get_parameter_or<double>("radius", radius, 0.5);
-
-  double z;
-  node->get_parameter_or<double>("z", z, 0.5);
 
   double focal_length;
   node->get_parameter_or<double>("focal_length", focal_length, 550.0);
@@ -109,44 +85,32 @@ int main(int argc, char** argv)
   sim.add(*mesh_ptr, Eigen::Isometry3d::Identity());
 
   // In the main (rendering) thread, begin orbiting...
-  const auto start = std::chrono::steady_clock::now();
-
   rclcpp::Rate rate(framerate);
 
   while (rclcpp::ok())
   {
-    double dt = std::chrono::duration<double>(std::chrono::steady_clock::now() - start).count();
-
-    Eigen::Vector3d camera_pos(radius * cos(dt * orbit_speed), radius * sin(dt * orbit_speed), z);
-
-    Eigen::Vector3d look_at(0, 0, 0);
-
-    const auto pose = lookat(camera_pos, look_at, Eigen::Vector3d(0, 0, 1));
-
-    const auto depth_img = sim.render(pose);
-
-    cv::Mat cv_img;
-    gl_depth_sim::toCvImage16u(depth_img, cv_img);
-
-    cv_bridge::CvImage image;
-    image.header.stamp = node->now();
-    image.header.frame_id = camera_frame;
-    image.encoding = sensor_msgs::image_encodings::TYPE_16UC1;
-    image.image = cv_img;
-    image_pub.publish(image.toImageMsg());
-
-    geometry_msgs::msg::TransformStamped transform_stamped;
-    transform_stamped.transform.translation.x = pose.translation().x();
-    transform_stamped.transform.translation.y = pose.translation().y();
-    transform_stamped.transform.translation.z = pose.translation().z();
-    transform_stamped.transform.rotation = tf2::toMsg(Eigen::Quaternion<double>(pose.linear()));
-
-    transform_stamped.header.stamp = image.header.stamp;
-    transform_stamped.header.frame_id = base_frame;
-    transform_stamped.child_frame_id = camera_frame;
-    broadcaster.sendTransform(transform_stamped);
-
     rate.sleep();
+    try
+    {
+      geometry_msgs::msg::TransformStamped tf_camera_st = buffer.lookupTransform(base_frame, camera_frame, tf2::TimePointZero);
+      const auto pose = tf2::transformToEigen(tf_camera_st);
+      const auto depth_img = sim.render(pose);
+
+      cv::Mat cv_img;
+      gl_depth_sim::toCvImage16u(depth_img, cv_img);
+
+      cv_bridge::CvImage image;
+      image.header.stamp = tf_camera_st.header.stamp;
+      image.header.frame_id = tf_camera_st.child_frame_id;
+      image.encoding = sensor_msgs::image_encodings::TYPE_16UC1;
+      image.image = cv_img;
+      image_pub.publish(image.toImageMsg());
+    }
+    catch (tf2::LookupException &e)
+    {
+      RCLCPP_ERROR_STREAM(node->get_logger(), e.what());
+    }
+
     rclcpp::spin_some(node);
   }
